Enzo src-directory reorganization: following some of James's suggestions part 4 - move some problem initializers that are just used for testing into a subdirectory called ``tests``
