+++ src/Enzo/_enzo.hpp
@@ -189,14 +189,14 @@ extern "C" {
 #include "initial/EnzoInitialCloud.hpp"
 #include "initial/EnzoInitialCollapse.hpp"
 #include "initial/EnzoInitialCosmology.hpp"
-#include "initial/EnzoInitialFeedbackTest.hpp"
-#include "initial/EnzoInitialGrackleTest.hpp"
+#include "tests/EnzoInitialFeedbackTest.hpp"
+#include "tests/EnzoInitialGrackleTest.hpp"
 #include "io/EnzoInitialHdf5.hpp"
 #include "initial/EnzoInitialImplosion2.hpp"
 #include "initial/EnzoInitialInclinedWave.hpp"
 #include "io/EnzoInitialMusic.hpp"
 #include "obsolete/EnzoInitialPm.hpp"
-#include "initial/EnzoInitialPpmlTest.hpp"
+#include "tests/EnzoInitialPpmlTest.hpp"
 #include "initial/EnzoInitialSedovArray2.hpp"
 #include "initial/EnzoInitialSedovArray3.hpp"
 #include "initial/EnzoInitialSedovRandom.hpp"
@@ -205,10 +205,10 @@ extern "C" {
 #include "initial/EnzoInitialTurbulence.hpp"
 #include "initial/EnzoInitialIsolatedGalaxy.hpp"
 #include "initial/EnzoInitialBurkertBodenheimer.hpp"
-#include "initial/EnzoInitialMergeSinksTest.hpp"
-#include "initial/EnzoInitialAccretionTest.hpp"
+#include "tests/EnzoInitialMergeSinksTest.hpp"
+#include "tests/EnzoInitialAccretionTest.hpp"
 #include "initial/EnzoInitialShuCollapse.hpp"
-#include "initial/EnzoInitialBBTest.hpp"
+#include "tests/EnzoInitialBBTest.hpp"
 
 #include "mesh/EnzoRefineShock.hpp"
 #include "mesh/EnzoRefineParticleMass.hpp"
